pyproject.toml
[project]
name = "geovision-lulc"
version = "0.1.0"
description = "Deep learning system for Land Use Land Cover semantic segmentation using Sentinel-2 imagery"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.4.0",
    "rasterio>=1.3.9",
    "geopandas>=0.14.3",
    "shapely>=2.0.3",
    "albumentations>=1.4.0",
    "transformers>=4.38.0",
    "timm>=0.9.16",
    "streamlit>=1.32.0",
    "matplotlib>=3.8.3",
    "seaborn>=0.13.2",
    "tqdm>=4.66.2",
    "pyyaml>=6.0.1",
    "grad-cam>=1.5.0",
    "einops>=0.7.0",
    "opencv-python-headless>=4.9.0",
    "tabulate>=0.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "black>=24.2.0",
    "isort>=5.13.2",
    "ruff>=0.3.0",
    "jupyter>=1.0.0"
]
